engines: rename the Apple engine coreml -> neurt, and give it a working LLM slot
The engine that serves Apple is now named for the runtime that implements it —
NeuRT, the Apple half of the sibling neurun repo — instead of for Apple's
framework. That name was already wrong twice over: it collided with the L1
runtime `coreml` (runtimes/coreml), which is a different thing in a different
registry, and it described a framework rather than an identity, so it could not
survive the engine growing a second modality.

ATOMIC BY NECESSITY. plugin_loader.cpp derives the entry symbol from the library
FILENAME (entry_symbol_from_path), so librunanywhere_neurt.dylib ->
rac_plugin_entry_neurt -> manifest .name "neurt" have to move in one change or a
dlopen host silently fails to install the vtable. RAC_ENGINE_ID_NEURT moves with
them, because model_lifecycle pins on that string; a drift there would not error,
it would quietly route every ANE model to whichever engine won on priority.
test_plugin_entry_neurt now asserts the id and the manifest name are equal.

THE LLM SLOT NOW ACTUALLY LINKS. It never had. The engine referenced
g_neurt_llm_ops, but that symbol was never emitted (const at namespace scope is
internal linkage even inside extern "C" — fixed on the neurun side), and nothing
here linked an executable against the engine, so librac_backend_neurt.a built
green while the ANE path was empty. Worse, the CMake only compiled the op-table
adapter and never linked neurt_core behind it, so even with the symbol present
the first real link failed on neurt::Generator::load. Both are fixed: the engine
derives neurt_core's source list from neurun's own CMakeLists (no second copy to
drift), and test_plugin_entry_neurt asserts llm_ops is non-null. The engine now
advertises GENERATE_TEXT unconditionally, so `rcli backends --json` reports
"neurt" with both diffusion and generate_text.

APPLE CODE MOVES OUT. rac_diffusion_coreml.{h,mm} go to neurun's
NeuRT/src/sdk/, next to rac_llm_ops_neurt.cpp. Consequence, stated plainly: the
neurun checkout is now a HARD requirement of this engine. NEURT_ROOT missing is
a FATAL_ERROR, not the silent degrade it used to be — an engine that registers,
claims Apple availability and then serves nothing is worse than a build error.
Use -DRAC_BACKEND_NEURT=OFF if you do not have the checkout.

TWO LATENT BUGS FIXED IN PASSING. The Electron addon and the Python module both
declared and called rac_backend_coreml_register(), which has never existed
anywhere in this tree — a hard link failure on any Apple configure that enabled
either target. Both now register the plugin entry directly, the way rcli always
has.

PRIORITY IS DELIBERATELY UNCHANGED at 100, below mlx's 110. Raising it would
hand the ANE every unpinned GENERATE_TEXT request, including GGUF and
safetensors models this engine cannot open. ANE models arrive by name pin
(INFERENCE_FRAMEWORK_COREML -> RAC_ENGINE_ID_NEURT ->
rac_plugin_find_for_engine), which bypasses priority entirely; that path is now
covered by a test.

NOT RENAMED, on purpose: RAC_RUNTIME_COREML, RAC_RUNTIME_ANE,
RAC_MODEL_FORMAT_ID_COREML, INFERENCE_FRAMEWORK_COREML, runtimes/coreml, the
rac_diffusion_coreml_* C ABI, and -framework CoreML. Those name Apple's
framework, not this engine.

RELEASE-ARTIFACT CONSEQUENCE: the xcframework is now RABackendNeuRT, so
Package.swift's remote binaryTarget points at RABackendNeuRT-ios-v<version>.zip,
which will not exist until the next release publishes it. Local development is
unaffected (RUNANYWHERE_USE_LOCAL_NATIVES=1). release.yml's backends gate now
greps for "name":"neurt".

Verified on macOS: rcli links and reports the engine; test_plugin_entry_neurt
passes; librunanywhere_neurt.dylib builds under RAC_STATIC_PLUGINS=OFF and a
dlopen + dlsym of the filename-derived rac_plugin_entry_neurt returns a vtable;
DRY_RUN=1 build-core-xcframework.sh is coherent end to end; both Package.swift
manifests parse. Electron cannot be built on macOS (MSVC-gated) and the Python
native module was not built — both changes there are the same three-line
register swap.
