+++ CMakePresets.json
@@ -97,7 +97,7 @@
                 "RAC_BACKEND_MLX": "ON",
                 "RAC_BACKEND_SHERPA": "ON",
                 "RAC_BACKEND_ONNX": "ON",
-                "RAC_BACKEND_COREML": "ON",
+                "RAC_BACKEND_NEURT": "ON",
                 "RAC_RUNTIME_ONNXRT": "ON",
                 "RAC_RUNTIME_COREML": "ON",
                 "RAC_BACKEND_RAG": "ON",
@@ -153,7 +153,7 @@
                 "RAC_BACKEND_SHERPA": "ON",
                 "RAC_BACKEND_ONNX": "ON",
                 "RAC_BACKEND_MLX": "OFF",
-                "RAC_BACKEND_COREML": "OFF",
+                "RAC_BACKEND_NEURT": "OFF",
                 "RAC_RUNTIME_ONNXRT": "ON",
                 "RAC_BACKEND_RAG": "OFF",
                 "RAC_BUILD_SERVER": "OFF",
