+++ engines/.clang-format
@@ -1,6 +1,6 @@
 # RunAnywhere Engines - Clang Format Configuration
 # Mirrors sdk/runanywhere-commons/.clang-format to keep engine plugin
-# sources (llamacpp, sherpa, onnx, qhexrt, coreml, cloud, common)
+# sources (llamacpp, sherpa, onnx, qhexrt, neurt, cloud, common)
 # on the same style as the C ABI/headers
 # they implement. Editing this file? Keep it in lockstep with
 # sdk/runanywhere-commons/.clang-format and runtimes/.clang-format.
