+++ examples/ios/RunAnywhereAI/README.md
@@ -62,7 +62,7 @@ Expected artifacts under `sdk/runanywhere-swift/Binaries/`:
 | `RABackendONNX.xcframework` · `RABackendSherpa.xcframework` | STT, TTS, VAD |
 | `onnxruntime.xcframework` · `onnx.xcframework` | ONNX Runtime for the two above |
 | `RABackendMLX.xcframework` · `RunAnywhereMLXRuntime.xcframework` · `RunAnywhereMLXMetal.xcframework` | Apple MLX (this app links `RunAnywhereMLX`) |
-| `RABackendCoreML.xcframework` | Core ML image generation |
+| `RABackendNeuRT.xcframework` | NeuRT — ANE text generation + Core ML image generation |
 
 All ten are produced by one run of the script. A short list is the usual cause of link errors on the MLX path, since this app's `Package.swift` depends on the `RunAnywhereMLX` product.
