+++ runtimes/AGENTS.md
@@ -8,7 +8,7 @@ This file provides guidance to AI coding assistants (Claude Code, Cursor, etc.) 
 
 A **runtime is a compute substrate / device** — the hardware (or vendor library treated as a device) that math actually runs on: **CPU**, **Apple Metal** (GPU), **Apple Core ML** (ANE/GPU/CPU chosen by Core ML), **ONNX Runtime as a library**, and future targets (CUDA, Vulkan, QNN, NNAPI, WebGPU). A runtime is **named by the device/framework** (`"cpu"`, `"metal"`, `"coreml"`, `"onnxrt"`), keyed by `rac_runtime_id_t` (`sdk/runanywhere-commons/include/rac/plugin/rac_primitive.h:83`).
 
-**A runtime is NOT an engine.** Engines (llamacpp, sherpa, onnx, coreml, qhexrt, cloud) serve **modalities** — they publish primitive ops (`llm_ops`, `stt_ops`, …) and are *clients* of one or more runtimes. Runtimes provide **compute**; they describe a device and, in exactly one case, host a session. The two registries are separate: engines register a `rac_engine_vtable_t` via `rac_plugin_register`; runtimes register a `rac_runtime_vtable_t` via `rac_runtime_register` (`sdk/runanywhere-commons/include/rac/plugin/rac_runtime_registry.h`).
+**A runtime is NOT an engine.** Engines (llamacpp, sherpa, onnx, neurt, qhexrt, cloud) serve **modalities** — they publish primitive ops (`llm_ops`, `stt_ops`, …) and are *clients* of one or more runtimes. Runtimes provide **compute**; they describe a device and, in exactly one case, host a session. The two registries are separate: engines register a `rac_engine_vtable_t` via `rac_plugin_register`; runtimes register a `rac_runtime_vtable_t` via `rac_runtime_register` (`sdk/runanywhere-commons/include/rac/plugin/rac_runtime_registry.h`).
 
 Promoting runtimes to first-class plugins lets multiple engines share one ORT `Ort::Env`, reuse one Core ML `MLModel` loader, and (eventually) allocate device buffers through one allocator per device instead of one per engine — see the header rationale in `sdk/runanywhere-commons/include/rac/plugin/rac_runtime_vtable.h:1-27`.
 
@@ -106,7 +106,7 @@ The vtable/registrar machinery is uniform across all three runtimes, but **how m
 
   > **`sherpa` does NOT use this runtime.** This is the most common misconception. The `sherpa` engine declares **`RAC_RUNTIME_CPU`** in `k_sherpa_runtimes[]` (`engines/sherpa/rac_plugin_entry_sherpa.cpp:47-48`), links the **sherpa-onnx** library + **raw `onnxruntime`** as static archives (`engines/sherpa/CMakeLists.txt:370-373`), and calls the sherpa-onnx C API directly. Its sources contain **zero** references to `runtime::onnxrt` / `rac_runtime_onnxrt`. So `onnxrt` = a thin ORT wrapper used by **one** engine (`onnx`); sherpa reaches ONNX Runtime transitively through sherpa-onnx, never through `runtimes/onnxrt`.
 
-- **coreml** — a capability runtime **plus a bag of MLModel loader helpers** (`runtimes/coreml/rac_runtime_coreml.mm:141-255`). Real consumer: the **`coreml` engine** (our diffusion pipeline, recently renamed from `diffusion-coreml`) via those loaders + `rac_coreml_runtime_require_available`, in `engines/coreml/rac_diffusion_coreml.mm`. **Same-name, different registry:** engine `coreml` (registers `rac_plugin_entry_coreml` / `RAC_STATIC_PLUGIN_REGISTER(coreml)`, `engines/coreml/rac_static_register_coreml.cpp:27`) and runtime `coreml` (registers `rac_runtime_entry_coreml`, `runtimes/coreml/rac_runtime_coreml.mm:262-266`) share the framework name but live in separate dirs/registries with distinct symbols — a clean illustration of **"an engine uses a same-named device runtime."**
+- **coreml** — a capability runtime **plus a bag of MLModel loader helpers** (`runtimes/coreml/rac_runtime_coreml.mm:141-255`). Real consumer: the **`neurt` engine** via those loaders + `rac_coreml_runtime_require_available`, in `NeuRT/src/sdk/rac_diffusion_coreml.mm` (that source lives in the sibling `neurun` repo; the SDK compiles it). **The name collision is gone:** the engine used to be called `coreml` too, and the pair was the standing example of "an engine uses a same-named device runtime". Since the engine became `neurt` (registering `rac_plugin_entry_neurt` / `RAC_STATIC_PLUGIN_REGISTER(neurt)`, `engines/neurt/rac_static_register_neurt.cpp`) only the runtime carries the framework name (`rac_runtime_entry_coreml`, `runtimes/coreml/rac_runtime_coreml.mm:262-266`). The separate dirs/registries/symbols were always what made it work; the shared name was never load-bearing.
 
 ---
 
@@ -147,7 +147,7 @@ An engine declares which device runtimes it needs in its manifest's `runtimes[]`
 
 **Pattern 2 — Engine uses a separate runtime as a library.** The engine declares the runtime and calls into the runtime's C++/helper API. `onnx` declares `RAC_RUNTIME_ONNXRT` (`engines/onnx/rac_plugin_entry_onnx.cpp:44,77`) and drives ORT via `runtime::onnxrt::Session`. **This is the runtime as a shared library, not a session host.**
 
-**Pattern 3 — Engine IS our own code on a device-runtime.** The engine is RunAnywhere code that runs on a device runtime via that runtime's loader/helpers. `coreml` declares `RAC_RUNTIME_COREML` (`engines/coreml/rac_plugin_entry_coreml.cpp:128,177`) and calls the Core ML loader helpers. (Contrast `cloud` engines: no runtime / pure HTTP — they declare no device runtime.)
+**Pattern 3 — Engine IS our own code on a device-runtime.** The engine is RunAnywhere code that runs on a device runtime via that runtime's loader/helpers. `neurt` declares `RAC_RUNTIME_COREML` + `RAC_RUNTIME_ANE` (`engines/neurt/rac_plugin_entry_neurt.cpp`) and calls the Core ML loader helpers. (Contrast `cloud` engines: no runtime / pure HTTP — they declare no device runtime.)
 
 ### THE RULE
