+++ sdk/runanywhere-cli/src/bootstrap.cpp
@@ -46,13 +46,13 @@
 #if defined(RCLI_HAS_MLX)
 #include "rac/backends/rac_mlx.h"
 #endif
-#if defined(RCLI_HAS_COREML)
-// The coreml engine (Apple-only, serves DIFFUSION) has no dedicated
-// rac_backend_coreml_register() fn; register its plugin entry directly. This
-// call also keeps the static rac_backend_coreml archive linked (references
-// rac_plugin_entry_coreml), mirroring how the other backends stay alive.
+#if defined(RCLI_HAS_NEURT)
+// The neurt engine (Apple-only: ANE LLM + CoreML diffusion) has no dedicated
+// rac_backend_neurt_register() fn; register its plugin entry directly. This
+// call also keeps the static rac_backend_neurt archive linked (references
+// rac_plugin_entry_neurt), mirroring how the other backends stay alive.
 #include "rac/plugin/rac_plugin_entry.h"
-#include "rac/plugin/rac_plugin_entry_coreml.h"
+#include "rac/plugin/rac_plugin_entry_neurt.h"
 #endif
 
 namespace rcli {
@@ -613,9 +613,9 @@ rac_result_t bootstrap(const GlobalOptions &options, Bootstrapped *out) {
           "warning: mlx backend requires MLX runtime callbacks; backend failed to register");
     }
 #endif
-#if defined(RCLI_HAS_COREML)
-    if (rac_plugin_register(rac_plugin_entry_coreml()) != RAC_SUCCESS) {
-      out::status_line("warning: coreml diffusion backend failed to register");
+#if defined(RCLI_HAS_NEURT)
+    if (rac_plugin_register(rac_plugin_entry_neurt()) != RAC_SUCCESS) {
+      out::status_line("warning: neurt (Apple Neural Engine) backend failed to register");
     }
 #endif
