+++ sdk/runanywhere-commons/scripts/build-ios.sh
@@ -62,11 +62,11 @@ xcframework_names=(
     RABackendONNX
     RABackendSherpa
     # Apple CoreML Stable-Diffusion engine. The root Package.swift's ONNXRuntime
-    # target has an UNCONDITIONAL .binaryTarget dependency on RABackendCoreMLBinary
-    # (RABackendCoreML-ios-v<version>.zip), so external SPM consumers 404 on
+    # target has an UNCONDITIONAL .binaryTarget dependency on RABackendNeuRTBinary
+    # (RABackendNeuRT-ios-v<version>.zip), so external SPM consumers 404 on
     # `swift package resolve` unless this archive is published. build-core-xcframework.sh
-    # already emits RABackendCoreML.xcframework (Apple-only, RAC_BACKEND_COREML=ON).
-    RABackendCoreML
+    # already emits RABackendNeuRT.xcframework (Apple-only, RAC_BACKEND_NEURT=ON).
+    RABackendNeuRT
     RABackendMLX
     RunAnywhereMLXRuntime
     RunAnywhereMLXMetal
