+++ sdk/runanywhere-react-native/scripts/package-sdk.sh
@@ -319,7 +319,7 @@ if [ -n "$NATIVES_FROM" ]; then
             RunAnywhereMLXRuntime \
             RunAnywhereMLXMetal
         stage_mlx_ios_resources
-        stage_ios "$RN_ROOT/packages/onnx" RABackendONNX RABackendSherpa RABackendCoreML
+        stage_ios "$RN_ROOT/packages/onnx" RABackendONNX RABackendSherpa RABackendNeuRT
     fi
 
     HAS_ANDROID=0
